Way too much code cleanup + Minor fixes related to squid&shark attack timing + fixed some step sounds

+++ src/main/java/its_meow/betteranimalsplus/common/entity/ai/LammerMoveHelper.java
@@ -31,7 +31,7 @@ public class LammerMoveHelper extends MovementController {
             }
             float yawAngle = (float)(MathHelper.atan2(diffZ, diffX) * (double)(180F / (float)Math.PI)) - 90.0F;
             this.parentEntity.rotationYaw = this.limitAngle(this.parentEntity.rotationYaw, yawAngle, 35.0F);
-            double xZDist = (double)MathHelper.sqrt(diffX * diffX + diffZ * diffZ);
+            double xZDist = MathHelper.sqrt(diffX * diffX + diffZ * diffZ);
             float pitchAngle = (float)(-(MathHelper.atan2(diffY, xZDist) * (double)(180F / (float)Math.PI)));
             this.parentEntity.rotationPitch = this.limitAngle(this.parentEntity.rotationPitch, pitchAngle, 35.0F);
         }
