+++ src/main/java/its_meow/betteranimalsplus/common/entity/ai/WaterfowlNavigator.java
@@ -37,7 +37,7 @@ public class WaterfowlNavigator extends GroundPathNavigator {
 
     @Override
     protected Vector3d getEntityPosition() {
-        return new Vector3d(this.entity.getPosX(), (double) this.getPathablePosY(), this.entity.getPosZ());
+        return new Vector3d(this.entity.getPosX(), this.getPathablePosY(), this.entity.getPosZ());
     }
 
     @Override
@@ -77,12 +77,12 @@ public class WaterfowlNavigator extends GroundPathNavigator {
     private int getPathablePosY() {
         if(this.entity.isInWater() && this.getCanSwim()) {
             int i = MathHelper.floor(this.entity.getBoundingBox().minY);
-            Block block = this.world.getBlockState(new BlockPos(this.entity.getPosX(), (double) i, this.entity.getPosZ())).getBlock();
+            Block block = this.world.getBlockState(new BlockPos(this.entity.getPosX(), i, this.entity.getPosZ())).getBlock();
             int j = 0;
 
             while(block == Blocks.WATER) {
                 ++i;
-                block = this.world.getBlockState(new BlockPos(this.entity.getPosX(), (double) i, this.entity.getPosZ())).getBlock();
+                block = this.world.getBlockState(new BlockPos(this.entity.getPosX(), i, this.entity.getPosZ())).getBlock();
                 ++j;
                 if(j > 16) {
                     return MathHelper.floor(this.entity.getBoundingBox().minY);
